Refactor MINAR code to allow more testing
Currently, the implementation details of MINAR (including internal
classes) are hidden inside source/minar.cpp. While that in itself
makes a lot of sense, it prevents some testing that can be done
using MINAR's internal components (and implementation details).
This commit takes one of those internal components (CallbackNode)
and puts it under a special header directory (minar-internal-headers).
Regular code wouldn't need to touch this, while tests that are
knowingly using MINAR's implementation details can include headers
in this directory and use them. It's expected that more such headers
will land in minar-internal-headers/ in the future, for the same
reason.

minar/trace.h is there just because it's needed from both minar.cpp
and CallbackNode.h (plus, it makes the tracing macros a little easier
to find).
